rasp_server/requirements.txt
flask
flask-socketio
eventlet
wiringpi
pyserial
